refactor(api): extract shared test fixtures + cleanup helper
Pulls `mockIndexAndSpecs`, the standard `afterEach` cleanup, and the
canonical Jobs/Mutate spec YAMLs into `__tests__/fixtures.ts`. The
three `api/*` test files (list, spec, call — added in the upstack PRs)
all reach for the same shapes; this is the first step toward each test
file containing only what it actually asserts.

Extends the ESLint dep-exemption to cover any `**/__tests__/**/*.ts`
file (was: only the legacy `test/__fixtures__/` dir) so colocated test
helpers can import nock/vitest like `*.test.ts` files do.
